Mean-zero'd the signals. Fixed end of convolution.
Had forgotten to find the avg then minus this from the signals,
now fixed.

Figured out why the last few calculations of the convolution
were giving wrong values, corrected with the help of printf
statements and now it is producing the same values as a matlab
convolution does, which is good enough for me.